+++ test/NameLookup/Inputs/members_B.swift
@@ -1,4 +1,4 @@
-import extensions_A
+import members_A
 
 
 extension X {
